backend/requirements.txt
# FastAPI Backend Dependencies for Sony Interior
# Phase 9 - Backend Setup

# Web Framework
fastapi==0.115.12
uvicorn[standard]==0.34.0

# AI Agent SDK and LLM Routing
openai-agents==0.1.0
litellm==1.58.5

# Database
asyncpg==0.30.0
pgvector==0.3.8

# Embeddings
sentence-transformers==3.4.1

# Data Validation
pydantic==2.10.6
pydantic-settings==2.7.2

# Environment Variables
python-dotenv==1.0.1

# HTTP Client
httpx==0.28.1

# Form Data
python-multipart==0.0.20

# CORS and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
python-dateutil==2.9.0
